GptQaDemo/requirements.txt
chromadb
pandas
openai
langchain
unstructured[local-inference]
gradio
python-dotenv
BeautifulSoup4